web/requirements.txt
azure-core==1.18.0
azure-cosmos==4.2.0
bson==0.5.10
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fire==0.4.0
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-WTF==0.15.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
oauthlib==3.1.1
pycparser==2.20
pymongo==3.12.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
termcolor==1.1.0
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
